+++ src/main/java/org/tokenscript/eip712/Eip712ExternalData.java
@@ -4,12 +4,12 @@ public class Eip712ExternalData {
   private String signatureInHex;
   // jsonRPC and chainID are needed to be compliant with EIP712 https://eips.ethereum.org/EIPS/eip-712
   private String jsonRpc;
-  private int chainId;
+  private long chainId;
   private String jsonSigned;
 
   public Eip712ExternalData() {}
 
-  public Eip712ExternalData(String signatureInHex, String jsonRpc, int chainId, String jsonSigned) {
+  public Eip712ExternalData(String signatureInHex, String jsonRpc, long chainId, String jsonSigned) {
     this.signatureInHex = signatureInHex;
     this.jsonRpc = jsonRpc;
     this.chainId = chainId;
@@ -32,11 +32,11 @@ public class Eip712ExternalData {
     this.jsonRpc = jsonRpc;
   }
 
-  public int getChainId() {
+  public long getChainId() {
     return chainId;
   }
 
-  public void setChainId(int chainId) {
+  public void setChainId(long chainId) {
     this.chainId = chainId;
   }
